[update] refs #47 (auth)ディレクトリに移動
 ディレクトリ構成改善のため
